requirements.txt
fastapi
uvicorn
psycopg2-binary
sentence-transformers
numpy
pandas
openpyxl
requests
python-multipart
pymupdf